Assignment 2/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-classic>=1.0.0
google-generativeai>=0.8.0
wikipedia>=1.4.0
duckduckgo-search>=6.0.0
ddgs>=9.0.0
python-dotenv>=1.0.0
markdown
